v2x_vlm/requirements.txt
torch
torchvision
transformers
datasets
numpy
pandas
matplotlib
seaborn
pyyaml
tqdm
scikit-learn
scipy
pillow
openai
